Move simple boundary factory in with the others
