+++ pyiron_continuum/fenics/job/elastic.py
@@ -5,18 +5,17 @@
 """
 A job class for FEM linear elasticity with [fenics](https://fenicsproject.org/pub/tutorial/html/._ftut1008.html).
 """
-from pyiron_base import ImportAlarm, HasStorage
+from pyiron_base import ImportAlarm
 with ImportAlarm(
         'fenics functionality requires the `fenics`, `mshr` modules (and their dependencies) specified as extra'
         'requirements. Please install it and try again.'
 ) as fenics_alarm:
     import fenics as FEN
     from ufl import nabla_div as ufl_nabla_div
-    from fenics import Constant, Expression, near
 
 from pyiron_continuum.fenics.job.generic import Fenics
 from pyiron_continuum.fenics.plot import Plot
-from pyiron_continuum.fenics.factory import SolverConfig, StringInputParser
+from pyiron_continuum.fenics.factory import SolverConfig, SimpleBoundaries
 
 __author__ = "Liam Huber"
 __copyright__ = (
@@ -28,23 +27,6 @@ __maintainer__ = "Liam Huber"
 __email__ = "huber@mpie.de"
 __status__ = "development"
 __date__ = "Dec 26, 2020"
-
-
-class SimpleBoundaries(HasStorage):
-    def __init__(self):
-        super().__init__()
-        self.storage.pairs = []
-
-    def add(self, value, condition):
-        StringInputParser(value)
-        StringInputParser(condition)
-        self.storage.pairs.append((value, condition))
-
-    def list(self):
-        return self.storage.pairs
-
-    def clear(self):
-        self.storage.pairs = []
 
 
 class FenicsLinearElastic(Fenics):
@@ -84,15 +66,6 @@ class FenicsLinearElastic(Fenics):
             self._solver = ElasticSolver(job=self)
         return self._solver
 
-    def _instantiate_bcs(self):
-        for (v, w) in self.input.boundaries.list():
-            def boundary_func(x, on_boundary):
-                try:
-                    return on_boundary and eval(w)
-                except Exception as err_msg:
-                    print(err_msg)
-            self.domain._bcs.append(FEN.DirichletBC(self.solver.V, eval(v), boundary_func))
-
     def _build_domain(self):
         self.domain.mesh.regular.box(
             (0, 0, 0),
@@ -101,10 +74,10 @@ class FenicsLinearElastic(Fenics):
             self.input.mesh_resolution,
             self.input.mesh_resolution
         )
+        self.domain._bcs = self.input.boundaries.to_fenics(self.solver.V)
 
     def validate_ready_to_run(self):
         self._build_domain()
-        self._instantiate_bcs()
         self.solver.set_sides_eq()
         super().validate_ready_to_run()
